telegram_bot/requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
redis==5.2.1